requirements.txt
streamlit>=1.26
pandas>=2.0
plotly>=5.17.0
kaleido>=0.2.1
fpdf2==2.7.8
matplotlib>=3.8.0